Bugfix of click event handling in tab panes
Bug was introduced in commit 1af5bf24d (New feature: TabPane
can be focused and navigated by keyboard) and affected
Internet Explorer 8. Tabs had to be clicked twice to change.

Fixed by avoiding to update the tab pane header (which removes and reattaches the click
listener which leads to losing the click event (?)) in the blur event handler.