Change Overflow handling of tab pane to support absolute positioning ...
within a tab. By specifying 'overflow:auto' on the content container
div instead of each content div, elements that are positioned absolutely
within a tab are now positioned and scroll properly.

In particular, this fixes an issue with echolot's JQuery-based
datepicker component that when opened on an input element in a tab would
end up in wrong positions when the user scrolled down in the tab.